+++ stories/DataTable/Example.js
@@ -37,9 +37,9 @@ class DataTableExample extends React.Component {
           infiniteScroll
           itemsPerPage={20}
           columns={[
-            {title: 'Row Number', render: (row, rowNum) => '#' + (rowNum + 1), width: '20%', minWidth: '75px', important: true},
-            {title: 'First Name', render: row => <span>{row.firstName}</span>, width: '40%', minWidth: '100px'},
-            {title: 'Last Name', render: row => <span>{row.lastName}</span>, width: '40%', minWidth: '100px'}
+              {title: 'Row Number', render: (row, rowNum) => '#' + (rowNum + 1), width: '20%', minWidth: '75px', important: true},
+              {title: 'First Name', render: row => <span>{row.firstName}</span>, width: '40%', minWidth: '100px'},
+              {title: 'Last Name', render: row => <span>{row.lastName}</span>, width: '40%', minWidth: '100px'}
           ]}
           />
       </div>