+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Types of changes:
 ## 4.1.2 - 2018-06-12
 ### Added
 - `<RichTextArea/>` - add a flag to generate absolute paths url links [#1746](https://github.com/wix/wix-style-react/pull/1746)
+- `<DataTable/>` - add info icon with tooltip to table header [#1770](https://github.com/wix/wix-style-react/pull/1770)
 
 ### Fixed
 - `<TimePicker/>` - Prevent typing letters [#1751](https://github.com/wix/wix-style-react/pull/1751)